refactor : 컴포넌트 관련 코드들을 components 폴더로 모음
- 컴포넌트 관련된 코드들을 components 폴더로 모음
