+++ public/js/components/carousel.js
@@ -1,4 +1,4 @@
-import { $ } from './utils.js';
+import { $ } from '../utils.js';
 
 const HOST = 'http://localhost:8000';
 const PRODUCT_NUMBER_IN_ONE_LINE = 7;
