+++ public/js/components/recent-product.js
@@ -1,10 +1,9 @@
-import { $ } from './utils.js';
+import { $ } from '../utils.js';
 
 class RecentProduct {
   constructor() {
     this.recentProductPopUp = $('.recent-product__pop-up');
   }
-  // ㅅ
 
   init() {}
 }
